Convert to a survival_curves object

as_survival_curves() is a generic function for converting objects to the survival_curves class. There is currently only a method for a data.frame.

as_survival_curves(x, ...)

# S3 method for data.frame
as_survival_curves(x, time = "time", id = "id", survival = "survival", ...)

Arguments

x
An object of the appropriate class.

...
Additional arguments to pass to survival_curves().

time
The name of the column containing the times a which survival probabilities were computed.

id
The name of the column containing an indentifier for each individual.

survival
The name of the column containing survival probabilities.

Value

A survival_curves object.

See also

Examples

# Create a data frame storing survival curves
N <- 3 # Number of individuals
times <- rep(seq(0, 30, 1/12))
rates = runif(N, .2, .5)
sc_df <- data.frame(
  id = rep(1:N, each = length(times)),
  time = rep(times, times = length(rates)),
  rate = rep(rates, each = length(times))
)
sc_df$survival <- 1 - stats::pexp(sc_df$time, sc_df$rate)

# Convert the data frame to a "survival_curves" object
as_survival_curves(sc_df)

#>             time id     survival
#> 1     0.00000000  1 1.000000e+00
#> 2     0.08333333  1 9.814881e-01
#> 3     0.16666667  1 9.633188e-01
#> 4     0.25000000  1 9.454859e-01
#> 5     0.33333333  1 9.279832e-01
#> 6     0.41666667  1 9.108044e-01
#> 7     0.50000000  1 8.939437e-01
#> 8     0.58333333  1 8.773950e-01
#> 9     0.66666667  1 8.611528e-01
#> 10    0.75000000  1 8.452112e-01
#> 11    0.83333333  1 8.295647e-01
#> 12    0.91666667  1 8.142078e-01
#> 13    1.00000000  1 7.991353e-01
#> 14    1.08333333  1 7.843417e-01
#> 15    1.16666667  1 7.698221e-01
#> 16    1.25000000  1 7.555712e-01
#> 17    1.33333333  1 7.415841e-01
#> 18    1.41666667  1 7.278559e-01
#> 19    1.50000000  1 7.143819e-01
#> 20    1.58333333  1 7.011573e-01
#> 21    1.66666667  1 6.881776e-01
#> 22    1.75000000  1 6.754381e-01
#> 23    1.83333333  1 6.629344e-01
#> 24    1.91666667  1 6.506622e-01
#> 25    2.00000000  1 6.386172e-01
#> 26    2.08333333  1 6.267952e-01
#> 27    2.16666667  1 6.151920e-01
#> 28    2.25000000  1 6.038036e-01
#> 29    2.33333333  1 5.926260e-01
#> 30    2.41666667  1 5.816554e-01
#> 31    2.50000000  1 5.708878e-01
#> 32    2.58333333  1 5.603196e-01
#> 33    2.66666667  1 5.499470e-01
#> 34    2.75000000  1 5.397664e-01
#> 35    2.83333333  1 5.297743e-01
#> 36    2.91666667  1 5.199671e-01
#> 37    3.00000000  1 5.103415e-01
#> 38    3.08333333  1 5.008941e-01
#> 39    3.16666667  1 4.916216e-01
#> 40    3.25000000  1 4.825207e-01
#> 41    3.33333333  1 4.735883e-01
#> 42    3.41666667  1 4.648213e-01
#> 43    3.50000000  1 4.562166e-01
#> 44    3.58333333  1 4.477711e-01
#> 45    3.66666667  1 4.394820e-01
#> 46    3.75000000  1 4.313464e-01
#> 47    3.83333333  1 4.233613e-01
#> 48    3.91666667  1 4.155241e-01
#> 49    4.00000000  1 4.078319e-01
#> 50    4.08333333  1 4.002822e-01
#> 51    4.16666667  1 3.928722e-01
#> 52    4.25000000  1 3.855993e-01
#> 53    4.33333333  1 3.784612e-01
#> 54    4.41666667  1 3.714551e-01
#> 55    4.50000000  1 3.645788e-01
#> 56    4.58333333  1 3.578297e-01
#> 57    4.66666667  1 3.512056e-01
#> 58    4.75000000  1 3.447041e-01
#> 59    4.83333333  1 3.383230e-01
#> 60    4.91666667  1 3.320599e-01
#> 61    5.00000000  1 3.259129e-01
#> 62    5.08333333  1 3.198796e-01
#> 63    5.16666667  1 3.139580e-01
#> 64    5.25000000  1 3.081460e-01
#> 65    5.33333333  1 3.024417e-01
#> 66    5.41666667  1 2.968429e-01
#> 67    5.50000000  1 2.913477e-01
#> 68    5.58333333  1 2.859543e-01
#> 69    5.66666667  1 2.806608e-01
#> 70    5.75000000  1 2.754652e-01
#> 71    5.83333333  1 2.703658e-01
#> 72    5.91666667  1 2.653608e-01
#> 73    6.00000000  1 2.604485e-01
#> 74    6.08333333  1 2.556271e-01
#> 75    6.16666667  1 2.508949e-01
#> 76    6.25000000  1 2.462504e-01
#> 77    6.33333333  1 2.416918e-01
#> 78    6.41666667  1 2.372176e-01
#> 79    6.50000000  1 2.328263e-01
#> 80    6.58333333  1 2.285162e-01
#> 81    6.66666667  1 2.242859e-01
#> 82    6.75000000  1 2.201340e-01
#> 83    6.83333333  1 2.160589e-01
#> 84    6.91666667  1 2.120592e-01
#> 85    7.00000000  1 2.081336e-01
#> 86    7.08333333  1 2.042806e-01
#> 87    7.16666667  1 2.004990e-01
#> 88    7.25000000  1 1.967874e-01
#> 89    7.33333333  1 1.931444e-01
#> 90    7.41666667  1 1.895690e-01
#> 91    7.50000000  1 1.860597e-01
#> 92    7.58333333  1 1.826154e-01
#> 93    7.66666667  1 1.792348e-01
#> 94    7.75000000  1 1.759168e-01
#> 95    7.83333333  1 1.726602e-01
#> 96    7.91666667  1 1.694640e-01
#> 97    8.00000000  1 1.663269e-01
#> 98    8.08333333  1 1.632478e-01
#> 99    8.16666667  1 1.602258e-01
#> 100   8.25000000  1 1.572597e-01
#> 101   8.33333333  1 1.543485e-01
#> 102   8.41666667  1 1.514912e-01
#> 103   8.50000000  1 1.486869e-01
#> 104   8.58333333  1 1.459344e-01
#> 105   8.66666667  1 1.432328e-01
#> 106   8.75000000  1 1.405813e-01
#> 107   8.83333333  1 1.379789e-01
#> 108   8.91666667  1 1.354246e-01
#> 109   9.00000000  1 1.329177e-01
#> 110   9.08333333  1 1.304571e-01
#> 111   9.16666667  1 1.280421e-01
#> 112   9.25000000  1 1.256718e-01
#> 113   9.33333333  1 1.233454e-01
#> 114   9.41666667  1 1.210620e-01
#> 115   9.50000000  1 1.188209e-01
#> 116   9.58333333  1 1.166213e-01
#> 117   9.66666667  1 1.144624e-01
#> 118   9.75000000  1 1.123435e-01
#> 119   9.83333333  1 1.102638e-01
#> 120   9.91666667  1 1.082226e-01
#> 121  10.00000000  1 1.062192e-01
#> 122  10.08333333  1 1.042529e-01
#> 123  10.16666667  1 1.023230e-01
#> 124  10.25000000  1 1.004288e-01
#> 125  10.33333333  1 9.856963e-02
#> 126  10.41666667  1 9.674491e-02
#> 127  10.50000000  1 9.495398e-02
#> 128  10.58333333  1 9.319620e-02
#> 129  10.66666667  1 9.147096e-02
#> 130  10.75000000  1 8.977765e-02
#> 131  10.83333333  1 8.811569e-02
#> 132  10.91666667  1 8.648450e-02
#> 133  11.00000000  1 8.488351e-02
#> 134  11.08333333  1 8.331215e-02
#> 135  11.16666667  1 8.176988e-02
#> 136  11.25000000  1 8.025616e-02
#> 137  11.33333333  1 7.877047e-02
#> 138  11.41666667  1 7.731227e-02
#> 139  11.50000000  1 7.588107e-02
#> 140  11.58333333  1 7.447637e-02
#> 141  11.66666667  1 7.309767e-02
#> 142  11.75000000  1 7.174449e-02
#> 143  11.83333333  1 7.041636e-02
#> 144  11.91666667  1 6.911282e-02
#> 145  12.00000000  1 6.783341e-02
#> 146  12.08333333  1 6.657768e-02
#> 147  12.16666667  1 6.534520e-02
#> 148  12.25000000  1 6.413553e-02
#> 149  12.33333333  1 6.294826e-02
#> 150  12.41666667  1 6.178297e-02
#> 151  12.50000000  1 6.063924e-02
#> 152  12.58333333  1 5.951669e-02
#> 153  12.66666667  1 5.841493e-02
#> 154  12.75000000  1 5.733355e-02
#> 155  12.83333333  1 5.627220e-02
#> 156  12.91666667  1 5.523049e-02
#> 157  13.00000000  1 5.420807e-02
#> 158  13.08333333  1 5.320457e-02
#> 159  13.16666667  1 5.221965e-02
#> 160  13.25000000  1 5.125297e-02
#> 161  13.33333333  1 5.030417e-02
#> 162  13.41666667  1 4.937295e-02
#> 163  13.50000000  1 4.845896e-02
#> 164  13.58333333  1 4.756189e-02
#> 165  13.66666667  1 4.668143e-02
#> 166  13.75000000  1 4.581726e-02
#> 167  13.83333333  1 4.496910e-02
#> 168  13.91666667  1 4.413663e-02
#> 169  14.00000000  1 4.331958e-02
#> 170  14.08333333  1 4.251765e-02
#> 171  14.16666667  1 4.173057e-02
#> 172  14.25000000  1 4.095805e-02
#> 173  14.33333333  1 4.019984e-02
#> 174  14.41666667  1 3.945566e-02
#> 175  14.50000000  1 3.872526e-02
#> 176  14.58333333  1 3.800838e-02
#> 177  14.66666667  1 3.730478e-02
#> 178  14.75000000  1 3.661419e-02
#> 179  14.83333333  1 3.593639e-02
#> 180  14.91666667  1 3.527114e-02
#> 181  15.00000000  1 3.461820e-02
#> 182  15.08333333  1 3.397735e-02
#> 183  15.16666667  1 3.334837e-02
#> 184  15.25000000  1 3.273102e-02
#> 185  15.33333333  1 3.212511e-02
#> 186  15.41666667  1 3.153041e-02
#> 187  15.50000000  1 3.094672e-02
#> 188  15.58333333  1 3.037384e-02
#> 189  15.66666667  1 2.981156e-02
#> 190  15.75000000  1 2.925969e-02
#> 191  15.83333333  1 2.871804e-02
#> 192  15.91666667  1 2.818641e-02
#> 193  16.00000000  1 2.766463e-02
#> 194  16.08333333  1 2.715250e-02
#> 195  16.16666667  1 2.664986e-02
#> 196  16.25000000  1 2.615652e-02
#> 197  16.33333333  1 2.567231e-02
#> 198  16.41666667  1 2.519706e-02
#> 199  16.50000000  1 2.473062e-02
#> 200  16.58333333  1 2.427281e-02
#> 201  16.66666667  1 2.382347e-02
#> 202  16.75000000  1 2.338245e-02
#> 203  16.83333333  1 2.294960e-02
#> 204  16.91666667  1 2.252476e-02
#> 205  17.00000000  1 2.210778e-02
#> 206  17.08333333  1 2.169852e-02
#> 207  17.16666667  1 2.129684e-02
#> 208  17.25000000  1 2.090260e-02
#> 209  17.33333333  1 2.051565e-02
#> 210  17.41666667  1 2.013586e-02
#> 211  17.50000000  1 1.976311e-02
#> 212  17.58333333  1 1.939726e-02
#> 213  17.66666667  1 1.903818e-02
#> 214  17.75000000  1 1.868574e-02
#> 215  17.83333333  1 1.833983e-02
#> 216  17.91666667  1 1.800033e-02
#> 217  18.00000000  1 1.766711e-02
#> 218  18.08333333  1 1.734005e-02
#> 219  18.16666667  1 1.701906e-02
#> 220  18.25000000  1 1.670400e-02
#> 221  18.33333333  1 1.639478e-02
#> 222  18.41666667  1 1.609128e-02
#> 223  18.50000000  1 1.579340e-02
#> 224  18.58333333  1 1.550103e-02
#> 225  18.66666667  1 1.521408e-02
#> 226  18.75000000  1 1.493244e-02
#> 227  18.83333333  1 1.465601e-02
#> 228  18.91666667  1 1.438470e-02
#> 229  19.00000000  1 1.411841e-02
#> 230  19.08333333  1 1.385705e-02
#> 231  19.16666667  1 1.360053e-02
#> 232  19.25000000  1 1.334876e-02
#> 233  19.33333333  1 1.310165e-02
#> 234  19.41666667  1 1.285911e-02
#> 235  19.50000000  1 1.262106e-02
#> 236  19.58333333  1 1.238742e-02
#> 237  19.66666667  1 1.215811e-02
#> 238  19.75000000  1 1.193304e-02
#> 239  19.83333333  1 1.171213e-02
#> 240  19.91666667  1 1.149532e-02
#> 241  20.00000000  1 1.128252e-02
#> 242  20.08333333  1 1.107366e-02
#> 243  20.16666667  1 1.086866e-02
#> 244  20.25000000  1 1.066746e-02
#> 245  20.33333333  1 1.046999e-02
#> 246  20.41666667  1 1.027617e-02
#> 247  20.50000000  1 1.008594e-02
#> 248  20.58333333  1 9.899225e-03
#> 249  20.66666667  1 9.715972e-03
#> 250  20.75000000  1 9.536110e-03
#> 251  20.83333333  1 9.359578e-03
#> 252  20.91666667  1 9.186315e-03
#> 253  21.00000000  1 9.016258e-03
#> 254  21.08333333  1 8.849350e-03
#> 255  21.16666667  1 8.685531e-03
#> 256  21.25000000  1 8.524745e-03
#> 257  21.33333333  1 8.366936e-03
#> 258  21.41666667  1 8.212048e-03
#> 259  21.50000000  1 8.060027e-03
#> 260  21.58333333  1 7.910820e-03
#> 261  21.66666667  1 7.764376e-03
#> 262  21.75000000  1 7.620642e-03
#> 263  21.83333333  1 7.479569e-03
#> 264  21.91666667  1 7.341108e-03
#> 265  22.00000000  1 7.205210e-03
#> 266  22.08333333  1 7.071828e-03
#> 267  22.16666667  1 6.940914e-03
#> 268  22.25000000  1 6.812425e-03
#> 269  22.33333333  1 6.686314e-03
#> 270  22.41666667  1 6.562537e-03
#> 271  22.50000000  1 6.441052e-03
#> 272  22.58333333  1 6.321816e-03
#> 273  22.66666667  1 6.204786e-03
#> 274  22.75000000  1 6.089924e-03
#> 275  22.83333333  1 5.977188e-03
#> 276  22.91666667  1 5.866538e-03
#> 277  23.00000000  1 5.757937e-03
#> 278  23.08333333  1 5.651347e-03
#> 279  23.16666667  1 5.546730e-03
#> 280  23.25000000  1 5.444049e-03
#> 281  23.33333333  1 5.343269e-03
#> 282  23.41666667  1 5.244355e-03
#> 283  23.50000000  1 5.147272e-03
#> 284  23.58333333  1 5.051986e-03
#> 285  23.66666667  1 4.958464e-03
#> 286  23.75000000  1 4.866673e-03
#> 287  23.83333333  1 4.776582e-03
#> 288  23.91666667  1 4.688158e-03
#> 289  24.00000000  1 4.601371e-03
#> 290  24.08333333  1 4.516191e-03
#> 291  24.16666667  1 4.432587e-03
#> 292  24.25000000  1 4.350532e-03
#> 293  24.33333333  1 4.269995e-03
#> 294  24.41666667  1 4.190949e-03
#> 295  24.50000000  1 4.113366e-03
#> 296  24.58333333  1 4.037220e-03
#> 297  24.66666667  1 3.962483e-03
#> 298  24.75000000  1 3.889130e-03
#> 299  24.83333333  1 3.817135e-03
#> 300  24.91666667  1 3.746472e-03
#> 301  25.00000000  1 3.677118e-03
#> 302  25.08333333  1 3.609047e-03
#> 303  25.16666667  1 3.542237e-03
#> 304  25.25000000  1 3.476663e-03
#> 305  25.33333333  1 3.412303e-03
#> 306  25.41666667  1 3.349135e-03
#> 307  25.50000000  1 3.287136e-03
#> 308  25.58333333  1 3.226285e-03
#> 309  25.66666667  1 3.166560e-03
#> 310  25.75000000  1 3.107941e-03
#> 311  25.83333333  1 3.050407e-03
#> 312  25.91666667  1 2.993938e-03
#> 313  26.00000000  1 2.938515e-03
#> 314  26.08333333  1 2.884117e-03
#> 315  26.16666667  1 2.830726e-03
#> 316  26.25000000  1 2.778324e-03
#> 317  26.33333333  1 2.726892e-03
#> 318  26.41666667  1 2.676412e-03
#> 319  26.50000000  1 2.626866e-03
#> 320  26.58333333  1 2.578238e-03
#> 321  26.66666667  1 2.530510e-03
#> 322  26.75000000  1 2.483665e-03
#> 323  26.83333333  1 2.437688e-03
#> 324  26.91666667  1 2.392562e-03
#> 325  27.00000000  1 2.348271e-03
#> 326  27.08333333  1 2.304800e-03
#> 327  27.16666667  1 2.262133e-03
#> 328  27.25000000  1 2.220257e-03
#> 329  27.33333333  1 2.179156e-03
#> 330  27.41666667  1 2.138815e-03
#> 331  27.50000000  1 2.099222e-03
#> 332  27.58333333  1 2.060361e-03
#> 333  27.66666667  1 2.022220e-03
#> 334  27.75000000  1 1.984785e-03
#> 335  27.83333333  1 1.948042e-03
#> 336  27.91666667  1 1.911980e-03
#> 337  28.00000000  1 1.876586e-03
#> 338  28.08333333  1 1.841847e-03
#> 339  28.16666667  1 1.807751e-03
#> 340  28.25000000  1 1.774286e-03
#> 341  28.33333333  1 1.741440e-03
#> 342  28.41666667  1 1.709203e-03
#> 343  28.50000000  1 1.677562e-03
#> 344  28.58333333  1 1.646507e-03
#> 345  28.66666667  1 1.616027e-03
#> 346  28.75000000  1 1.586111e-03
#> 347  28.83333333  1 1.556749e-03
#> 348  28.91666667  1 1.527931e-03
#> 349  29.00000000  1 1.499646e-03
#> 350  29.08333333  1 1.471885e-03
#> 351  29.16666667  1 1.444637e-03
#> 352  29.25000000  1 1.417894e-03
#> 353  29.33333333  1 1.391646e-03
#> 354  29.41666667  1 1.365884e-03
#> 355  29.50000000  1 1.340599e-03
#> 356  29.58333333  1 1.315782e-03
#> 357  29.66666667  1 1.291424e-03
#> 358  29.75000000  1 1.267518e-03
#> 359  29.83333333  1 1.244053e-03
#> 360  29.91666667  1 1.221024e-03
#> 361  30.00000000  1 1.198420e-03
#> 362   0.00000000  2 1.000000e+00
#> 363   0.08333333  2 9.631703e-01
#> 364   0.16666667  2 9.276971e-01
#> 365   0.25000000  2 8.935303e-01
#> 366   0.33333333  2 8.606219e-01
#> 367   0.41666667  2 8.289255e-01
#> 368   0.50000000  2 7.983965e-01
#> 369   0.58333333  2 7.689918e-01
#> 370   0.66666667  2 7.406701e-01
#> 371   0.75000000  2 7.133915e-01
#> 372   0.83333333  2 6.871175e-01
#> 373   0.91666667  2 6.618112e-01
#> 374   1.00000000  2 6.374369e-01
#> 375   1.08333333  2 6.139604e-01
#> 376   1.16666667  2 5.913484e-01
#> 377   1.25000000  2 5.695693e-01
#> 378   1.33333333  2 5.485922e-01
#> 379   1.41666667  2 5.283878e-01
#> 380   1.50000000  2 5.089274e-01
#> 381   1.58333333  2 4.901838e-01
#> 382   1.66666667  2 4.721305e-01
#> 383   1.75000000  2 4.547421e-01
#> 384   1.83333333  2 4.379941e-01
#> 385   1.91666667  2 4.218629e-01
#> 386   2.00000000  2 4.063259e-01
#> 387   2.08333333  2 3.913610e-01
#> 388   2.16666667  2 3.769473e-01
#> 389   2.25000000  2 3.630645e-01
#> 390   2.33333333  2 3.496930e-01
#> 391   2.41666667  2 3.368139e-01
#> 392   2.50000000  2 3.244091e-01
#> 393   2.58333333  2 3.124613e-01
#> 394   2.66666667  2 3.009534e-01
#> 395   2.75000000  2 2.898694e-01
#> 396   2.83333333  2 2.791936e-01
#> 397   2.91666667  2 2.689110e-01
#> 398   3.00000000  2 2.590071e-01
#> 399   3.08333333  2 2.494680e-01
#> 400   3.16666667  2 2.402802e-01
#> 401   3.25000000  2 2.314307e-01
#> 402   3.33333333  2 2.229072e-01
#> 403   3.41666667  2 2.146976e-01
#> 404   3.50000000  2 2.067904e-01
#> 405   3.58333333  2 1.991744e-01
#> 406   3.66666667  2 1.918388e-01
#> 407   3.75000000  2 1.847735e-01
#> 408   3.83333333  2 1.779683e-01
#> 409   3.91666667  2 1.714138e-01
#> 410   4.00000000  2 1.651007e-01
#> 411   4.08333333  2 1.590201e-01
#> 412   4.16666667  2 1.531634e-01
#> 413   4.25000000  2 1.475225e-01
#> 414   4.33333333  2 1.420893e-01
#> 415   4.41666667  2 1.368562e-01
#> 416   4.50000000  2 1.318158e-01
#> 417   4.58333333  2 1.269611e-01
#> 418   4.66666667  2 1.222852e-01
#> 419   4.75000000  2 1.177814e-01
#> 420   4.83333333  2 1.134436e-01
#> 421   4.91666667  2 1.092655e-01
#> 422   5.00000000  2 1.052413e-01
#> 423   5.08333333  2 1.013653e-01
#> 424   5.16666667  2 9.763204e-02
#> 425   5.25000000  2 9.403629e-02
#> 426   5.33333333  2 9.057296e-02
#> 427   5.41666667  2 8.723719e-02
#> 428   5.50000000  2 8.402428e-02
#> 429   5.58333333  2 8.092969e-02
#> 430   5.66666667  2 7.794908e-02
#> 431   5.75000000  2 7.507824e-02
#> 432   5.83333333  2 7.231314e-02
#> 433   5.91666667  2 6.964987e-02
#> 434   6.00000000  2 6.708469e-02
#> 435   6.08333333  2 6.461398e-02
#> 436   6.16666667  2 6.223427e-02
#> 437   6.25000000  2 5.994220e-02
#> 438   6.33333333  2 5.773455e-02
#> 439   6.41666667  2 5.560821e-02
#> 440   6.50000000  2 5.356018e-02
#> 441   6.58333333  2 5.158758e-02
#> 442   6.66666667  2 4.968762e-02
#> 443   6.75000000  2 4.785765e-02
#> 444   6.83333333  2 4.609506e-02
#> 445   6.91666667  2 4.439740e-02
#> 446   7.00000000  2 4.276226e-02
#> 447   7.08333333  2 4.118734e-02
#> 448   7.16666667  2 3.967042e-02
#> 449   7.25000000  2 3.820938e-02
#> 450   7.33333333  2 3.680214e-02
#> 451   7.41666667  2 3.544673e-02
#> 452   7.50000000  2 3.414124e-02
#> 453   7.58333333  2 3.288383e-02
#> 454   7.66666667  2 3.167273e-02
#> 455   7.75000000  2 3.050623e-02
#> 456   7.83333333  2 2.938270e-02
#> 457   7.91666667  2 2.830054e-02
#> 458   8.00000000  2 2.725824e-02
#> 459   8.08333333  2 2.625433e-02
#> 460   8.16666667  2 2.528739e-02
#> 461   8.25000000  2 2.435607e-02
#> 462   8.33333333  2 2.345904e-02
#> 463   8.41666667  2 2.259505e-02
#> 464   8.50000000  2 2.176289e-02
#> 465   8.58333333  2 2.096137e-02
#> 466   8.66666667  2 2.018937e-02
#> 467   8.75000000  2 1.944580e-02
#> 468   8.83333333  2 1.872962e-02
#> 469   8.91666667  2 1.803981e-02
#> 470   9.00000000  2 1.737541e-02
#> 471   9.08333333  2 1.673548e-02
#> 472   9.16666667  2 1.611912e-02
#> 473   9.25000000  2 1.552546e-02
#> 474   9.33333333  2 1.495366e-02
#> 475   9.41666667  2 1.440292e-02
#> 476   9.50000000  2 1.387247e-02
#> 477   9.58333333  2 1.336155e-02
#> 478   9.66666667  2 1.286945e-02
#> 479   9.75000000  2 1.239547e-02
#> 480   9.83333333  2 1.193895e-02
#> 481   9.91666667  2 1.149924e-02
#> 482  10.00000000  2 1.107573e-02
#> 483  10.08333333  2 1.066781e-02
#> 484  10.16666667  2 1.027492e-02
#> 485  10.25000000  2 9.896500e-03
#> 486  10.33333333  2 9.532016e-03
#> 487  10.41666667  2 9.180955e-03
#> 488  10.50000000  2 8.842823e-03
#> 489  10.58333333  2 8.517145e-03
#> 490  10.66666667  2 8.203462e-03
#> 491  10.75000000  2 7.901331e-03
#> 492  10.83333333  2 7.610328e-03
#> 493  10.91666667  2 7.330042e-03
#> 494  11.00000000  2 7.060079e-03
#> 495  11.08333333  2 6.800059e-03
#> 496  11.16666667  2 6.549615e-03
#> 497  11.25000000  2 6.308395e-03
#> 498  11.33333333  2 6.076059e-03
#> 499  11.41666667  2 5.852280e-03
#> 500  11.50000000  2 5.636742e-03
#> 501  11.58333333  2 5.429143e-03
#> 502  11.66666667  2 5.229190e-03
#> 503  11.75000000  2 5.036600e-03
#> 504  11.83333333  2 4.851104e-03
#> 505  11.91666667  2 4.672440e-03
#> 506  12.00000000  2 4.500355e-03
#> 507  12.08333333  2 4.334609e-03
#> 508  12.16666667  2 4.174967e-03
#> 509  12.25000000  2 4.021204e-03
#> 510  12.33333333  2 3.873104e-03
#> 511  12.41666667  2 3.730459e-03
#> 512  12.50000000  2 3.593068e-03
#> 513  12.58333333  2 3.460736e-03
#> 514  12.66666667  2 3.333279e-03
#> 515  12.75000000  2 3.210515e-03
#> 516  12.83333333  2 3.092273e-03
#> 517  12.91666667  2 2.978386e-03
#> 518  13.00000000  2 2.868693e-03
#> 519  13.08333333  2 2.763040e-03
#> 520  13.16666667  2 2.661278e-03
#> 521  13.25000000  2 2.563264e-03
#> 522  13.33333333  2 2.468860e-03
#> 523  13.41666667  2 2.377933e-03
#> 524  13.50000000  2 2.290354e-03
#> 525  13.58333333  2 2.206001e-03
#> 526  13.66666667  2 2.124755e-03
#> 527  13.75000000  2 2.046501e-03
#> 528  13.83333333  2 1.971129e-03
#> 529  13.91666667  2 1.898533e-03
#> 530  14.00000000  2 1.828611e-03
#> 531  14.08333333  2 1.761264e-03
#> 532  14.16666667  2 1.696397e-03
#> 533  14.25000000  2 1.633919e-03
#> 534  14.33333333  2 1.573743e-03
#> 535  14.41666667  2 1.515782e-03
#> 536  14.50000000  2 1.459956e-03
#> 537  14.58333333  2 1.406187e-03
#> 538  14.66666667  2 1.354397e-03
#> 539  14.75000000  2 1.304515e-03
#> 540  14.83333333  2 1.256470e-03
#> 541  14.91666667  2 1.210195e-03
#> 542  15.00000000  2 1.165624e-03
#> 543  15.08333333  2 1.122695e-03
#> 544  15.16666667  2 1.081346e-03
#> 545  15.25000000  2 1.041520e-03
#> 546  15.33333333  2 1.003162e-03
#> 547  15.41666667  2 9.662155e-04
#> 548  15.50000000  2 9.306301e-04
#> 549  15.58333333  2 8.963554e-04
#> 550  15.66666667  2 8.633429e-04
#> 551  15.75000000  2 8.315463e-04
#> 552  15.83333333  2 8.009207e-04
#> 553  15.91666667  2 7.714231e-04
#> 554  16.00000000  2 7.430118e-04
#> 555  16.08333333  2 7.156470e-04
#> 556  16.16666667  2 6.892899e-04
#> 557  16.25000000  2 6.639036e-04
#> 558  16.33333333  2 6.394523e-04
#> 559  16.41666667  2 6.159015e-04
#> 560  16.50000000  2 5.932180e-04
#> 561  16.58333333  2 5.713700e-04
#> 562  16.66666667  2 5.503267e-04
#> 563  16.75000000  2 5.300583e-04
#> 564  16.83333333  2 5.105365e-04
#> 565  16.91666667  2 4.917336e-04
#> 566  17.00000000  2 4.736232e-04
#> 567  17.08333333  2 4.561798e-04
#> 568  17.16666667  2 4.393789e-04
#> 569  17.25000000  2 4.231967e-04
#> 570  17.33333333  2 4.076105e-04
#> 571  17.41666667  2 3.925984e-04
#> 572  17.50000000  2 3.781391e-04
#> 573  17.58333333  2 3.642124e-04
#> 574  17.66666667  2 3.507986e-04
#> 575  17.75000000  2 3.378788e-04
#> 576  17.83333333  2 3.254348e-04
#> 577  17.91666667  2 3.134492e-04
#> 578  18.00000000  2 3.019049e-04
#> 579  18.08333333  2 2.907859e-04
#> 580  18.16666667  2 2.800763e-04
#> 581  18.25000000  2 2.697612e-04
#> 582  18.33333333  2 2.598260e-04
#> 583  18.41666667  2 2.502567e-04
#> 584  18.50000000  2 2.410398e-04
#> 585  18.58333333  2 2.321624e-04
#> 586  18.66666667  2 2.236120e-04
#> 587  18.75000000  2 2.153764e-04
#> 588  18.83333333  2 2.074442e-04
#> 589  18.91666667  2 1.998041e-04
#> 590  19.00000000  2 1.924454e-04
#> 591  19.08333333  2 1.853577e-04
#> 592  19.16666667  2 1.785310e-04
#> 593  19.25000000  2 1.719558e-04
#> 594  19.33333333  2 1.656227e-04
#> 595  19.41666667  2 1.595229e-04
#> 596  19.50000000  2 1.536477e-04
#> 597  19.58333333  2 1.479889e-04
#> 598  19.66666667  2 1.425385e-04
#> 599  19.75000000  2 1.372889e-04
#> 600  19.83333333  2 1.322326e-04
#> 601  19.91666667  2 1.273625e-04
#> 602  20.00000000  2 1.226718e-04
#> 603  20.08333333  2 1.181538e-04
#> 604  20.16666667  2 1.138023e-04
#> 605  20.25000000  2 1.096110e-04
#> 606  20.33333333  2 1.055740e-04
#> 607  20.41666667  2 1.016858e-04
#> 608  20.50000000  2 9.794072e-05
#> 609  20.58333333  2 9.433360e-05
#> 610  20.66666667  2 9.085932e-05
#> 611  20.75000000  2 8.751300e-05
#> 612  20.83333333  2 8.428993e-05
#> 613  20.91666667  2 8.118556e-05
#> 614  21.00000000  2 7.819553e-05
#> 615  21.08333333  2 7.531561e-05
#> 616  21.16666667  2 7.254176e-05
#> 617  21.25000000  2 6.987007e-05
#> 618  21.33333333  2 6.729678e-05
#> 619  21.41666667  2 6.481827e-05
#> 620  21.50000000  2 6.243103e-05
#> 621  21.58333333  2 6.013172e-05
#> 622  21.66666667  2 5.791709e-05
#> 623  21.75000000  2 5.578402e-05
#> 624  21.83333333  2 5.372952e-05
#> 625  21.91666667  2 5.175068e-05
#> 626  22.00000000  2 4.984472e-05
#> 627  22.08333333  2 4.800895e-05
#> 628  22.16666667  2 4.624080e-05
#> 629  22.25000000  2 4.453777e-05
#> 630  22.33333333  2 4.289746e-05
#> 631  22.41666667  2 4.131756e-05
#> 632  22.50000000  2 3.979585e-05
#> 633  22.58333333  2 3.833018e-05
#> 634  22.66666667  2 3.691849e-05
#> 635  22.75000000  2 3.555880e-05
#> 636  22.83333333  2 3.424918e-05
#> 637  22.91666667  2 3.298779e-05
#> 638  23.00000000  2 3.177286e-05
#> 639  23.08333333  2 3.060268e-05
#> 640  23.16666667  2 2.947559e-05
#> 641  23.25000000  2 2.839002e-05
#> 642  23.33333333  2 2.734442e-05
#> 643  23.41666667  2 2.633734e-05
#> 644  23.50000000  2 2.536734e-05
#> 645  23.58333333  2 2.443307e-05
#> 646  23.66666667  2 2.353321e-05
#> 647  23.75000000  2 2.266649e-05
#> 648  23.83333333  2 2.183169e-05
#> 649  23.91666667  2 2.102764e-05
#> 650  24.00000000  2 2.025320e-05
#> 651  24.08333333  2 1.950728e-05
#> 652  24.16666667  2 1.878883e-05
#> 653  24.25000000  2 1.809685e-05
#> 654  24.33333333  2 1.743035e-05
#> 655  24.41666667  2 1.678839e-05
#> 656  24.50000000  2 1.617008e-05
#> 657  24.58333333  2 1.557454e-05
#> 658  24.66666667  2 1.500094e-05
#> 659  24.75000000  2 1.444846e-05
#> 660  24.83333333  2 1.391633e-05
#> 661  24.91666667  2 1.340379e-05
#> 662  25.00000000  2 1.291014e-05
#> 663  25.08333333  2 1.243466e-05
#> 664  25.16666667  2 1.197670e-05
#> 665  25.25000000  2 1.153560e-05
#> 666  25.33333333  2 1.111075e-05
#> 667  25.41666667  2 1.070154e-05
#> 668  25.50000000  2 1.030741e-05
#> 669  25.58333333  2 9.927789e-06
#> 670  25.66666667  2 9.562152e-06
#> 671  25.75000000  2 9.209982e-06
#> 672  25.83333333  2 8.870781e-06
#> 673  25.91666667  2 8.544073e-06
#> 674  26.00000000  2 8.229398e-06
#> 675  26.08333333  2 7.926312e-06
#> 676  26.16666667  2 7.634389e-06
#> 677  26.25000000  2 7.353217e-06
#> 678  26.33333333  2 7.082400e-06
#> 679  26.41666667  2 6.821558e-06
#> 680  26.50000000  2 6.570322e-06
#> 681  26.58333333  2 6.328340e-06
#> 682  26.66666667  2 6.095269e-06
#> 683  26.75000000  2 5.870783e-06
#> 684  26.83333333  2 5.654564e-06
#> 685  26.91666667  2 5.446308e-06
#> 686  27.00000000  2 5.245722e-06
#> 687  27.08333333  2 5.052524e-06
#> 688  27.16666667  2 4.866441e-06
#> 689  27.25000000  2 4.687212e-06
#> 690  27.33333333  2 4.514584e-06
#> 691  27.41666667  2 4.348313e-06
#> 692  27.50000000  2 4.188166e-06
#> 693  27.58333333  2 4.033918e-06
#> 694  27.66666667  2 3.885350e-06
#> 695  27.75000000  2 3.742254e-06
#> 696  27.83333333  2 3.604428e-06
#> 697  27.91666667  2 3.471678e-06
#> 698  28.00000000  2 3.343817e-06
#> 699  28.08333333  2 3.220666e-06
#> 700  28.16666667  2 3.102050e-06
#> 701  28.25000000  2 2.987802e-06
#> 702  28.33333333  2 2.877762e-06
#> 703  28.41666667  2 2.771775e-06
#> 704  28.50000000  2 2.669692e-06
#> 705  28.58333333  2 2.571368e-06
#> 706  28.66666667  2 2.476666e-06
#> 707  28.75000000  2 2.385451e-06
#> 708  28.83333333  2 2.297595e-06
#> 709  28.91666667  2 2.212976e-06
#> 710  29.00000000  2 2.131473e-06
#> 711  29.08333333  2 2.052971e-06
#> 712  29.16666667  2 1.977361e-06
#> 713  29.25000000  2 1.904536e-06
#> 714  29.33333333  2 1.834392e-06
#> 715  29.41666667  2 1.766832e-06
#> 716  29.50000000  2 1.701760e-06
#> 717  29.58333333  2 1.639085e-06
#> 718  29.66666667  2 1.578718e-06
#> 719  29.75000000  2 1.520574e-06
#> 720  29.83333333  2 1.464572e-06
#> 721  29.91666667  2 1.410633e-06
#> 722  30.00000000  2 1.358679e-06
#> 723   0.00000000  3 1.000000e+00
#> 724   0.08333333  3 9.688110e-01
#> 725   0.16666667  3 9.385948e-01
#> 726   0.25000000  3 9.093210e-01
#> 727   0.33333333  3 8.809603e-01
#> 728   0.41666667  3 8.534840e-01
#> 729   0.50000000  3 8.268648e-01
#> 730   0.58333333  3 8.010757e-01
#> 731   0.66666667  3 7.760910e-01
#> 732   0.75000000  3 7.518855e-01
#> 733   0.83333333  3 7.284350e-01
#> 734   0.91666667  3 7.057159e-01
#> 735   1.00000000  3 6.837053e-01
#> 736   1.08333333  3 6.623813e-01
#> 737   1.16666667  3 6.417223e-01
#> 738   1.25000000  3 6.217076e-01
#> 739   1.33333333  3 6.023172e-01
#> 740   1.41666667  3 5.835316e-01
#> 741   1.50000000  3 5.653318e-01
#> 742   1.58333333  3 5.476997e-01
#> 743   1.66666667  3 5.306175e-01
#> 744   1.75000000  3 5.140681e-01
#> 745   1.83333333  3 4.980349e-01
#> 746   1.91666667  3 4.825017e-01
#> 747   2.00000000  3 4.674530e-01
#> 748   2.08333333  3 4.528736e-01
#> 749   2.16666667  3 4.387489e-01
#> 750   2.25000000  3 4.250648e-01
#> 751   2.33333333  3 4.118075e-01
#> 752   2.41666667  3 3.989636e-01
#> 753   2.50000000  3 3.865204e-01
#> 754   2.58333333  3 3.744652e-01
#> 755   2.66666667  3 3.627860e-01
#> 756   2.75000000  3 3.514711e-01
#> 757   2.83333333  3 3.405091e-01
#> 758   2.91666667  3 3.298890e-01
#> 759   3.00000000  3 3.196001e-01
#> 760   3.08333333  3 3.096321e-01
#> 761   3.16666667  3 2.999750e-01
#> 762   3.25000000  3 2.906191e-01
#> 763   3.33333333  3 2.815550e-01
#> 764   3.41666667  3 2.727736e-01
#> 765   3.50000000  3 2.642660e-01
#> 766   3.58333333  3 2.560239e-01
#> 767   3.66666667  3 2.480387e-01
#> 768   3.75000000  3 2.403027e-01
#> 769   3.83333333  3 2.328079e-01
#> 770   3.91666667  3 2.255469e-01
#> 771   4.00000000  3 2.185123e-01
#> 772   4.08333333  3 2.116971e-01
#> 773   4.16666667  3 2.050945e-01
#> 774   4.25000000  3 1.986978e-01
#> 775   4.33333333  3 1.925006e-01
#> 776   4.41666667  3 1.864967e-01
#> 777   4.50000000  3 1.806801e-01
#> 778   4.58333333  3 1.750449e-01
#> 779   4.66666667  3 1.695854e-01
#> 780   4.75000000  3 1.642962e-01
#> 781   4.83333333  3 1.591720e-01
#> 782   4.91666667  3 1.542076e-01
#> 783   5.00000000  3 1.493980e-01
#> 784   5.08333333  3 1.447384e-01
#> 785   5.16666667  3 1.402242e-01
#> 786   5.25000000  3 1.358508e-01
#> 787   5.33333333  3 1.316137e-01
#> 788   5.41666667  3 1.275088e-01
#> 789   5.50000000  3 1.235319e-01
#> 790   5.58333333  3 1.196791e-01
#> 791   5.66666667  3 1.159464e-01
#> 792   5.75000000  3 1.123302e-01
#> 793   5.83333333  3 1.088267e-01
#> 794   5.91666667  3 1.054325e-01
#> 795   6.00000000  3 1.021442e-01
#> 796   6.08333333  3 9.895844e-02
#> 797   6.16666667  3 9.587203e-02
#> 798   6.25000000  3 9.288188e-02
#> 799   6.33333333  3 8.998499e-02
#> 800   6.41666667  3 8.717846e-02
#> 801   6.50000000  3 8.445945e-02
#> 802   6.58333333  3 8.182525e-02
#> 803   6.66666667  3 7.927320e-02
#> 804   6.75000000  3 7.680076e-02
#> 805   6.83333333  3 7.440542e-02
#> 806   6.91666667  3 7.208479e-02
#> 807   7.00000000  3 6.983654e-02
#> 808   7.08333333  3 6.765841e-02
#> 809   7.16666667  3 6.554822e-02
#> 810   7.25000000  3 6.350384e-02
#> 811   7.33333333  3 6.152322e-02
#> 812   7.41666667  3 5.960437e-02
#> 813   7.50000000  3 5.774538e-02
#> 814   7.58333333  3 5.594436e-02
#> 815   7.66666667  3 5.419951e-02
#> 816   7.75000000  3 5.250909e-02
#> 817   7.83333333  3 5.087138e-02
#> 818   7.91666667  3 4.928476e-02
#> 819   8.00000000  3 4.774762e-02
#> 820   8.08333333  3 4.625842e-02
#> 821   8.16666667  3 4.481567e-02
#> 822   8.25000000  3 4.341791e-02
#> 823   8.33333333  3 4.206375e-02
#> 824   8.41666667  3 4.075183e-02
#> 825   8.50000000  3 3.948082e-02
#> 826   8.58333333  3 3.824946e-02
#> 827   8.66666667  3 3.705650e-02
#> 828   8.75000000  3 3.590074e-02
#> 829   8.83333333  3 3.478103e-02
#> 830   8.91666667  3 3.369625e-02
#> 831   9.00000000  3 3.264530e-02
#> 832   9.08333333  3 3.162713e-02
#> 833   9.16666667  3 3.064071e-02
#> 834   9.25000000  3 2.968506e-02
#> 835   9.33333333  3 2.875921e-02
#> 836   9.41666667  3 2.786224e-02
#> 837   9.50000000  3 2.699325e-02
#> 838   9.58333333  3 2.615136e-02
#> 839   9.66666667  3 2.533572e-02
#> 840   9.75000000  3 2.454553e-02
#> 841   9.83333333  3 2.377998e-02
#> 842   9.91666667  3 2.303831e-02
#> 843  10.00000000  3 2.231977e-02
#> 844  10.08333333  3 2.162364e-02
#> 845  10.16666667  3 2.094922e-02
#> 846  10.25000000  3 2.029583e-02
#> 847  10.33333333  3 1.966283e-02
#> 848  10.41666667  3 1.904956e-02
#> 849  10.50000000  3 1.845543e-02
#> 850  10.58333333  3 1.787982e-02
#> 851  10.66666667  3 1.732217e-02
#> 852  10.75000000  3 1.678191e-02
#> 853  10.83333333  3 1.625850e-02
#> 854  10.91666667  3 1.575141e-02
#> 855  11.00000000  3 1.526014e-02
#> 856  11.08333333  3 1.478419e-02
#> 857  11.16666667  3 1.432309e-02
#> 858  11.25000000  3 1.387637e-02
#> 859  11.33333333  3 1.344358e-02
#> 860  11.41666667  3 1.302429e-02
#> 861  11.50000000  3 1.261807e-02
#> 862  11.58333333  3 1.222453e-02
#> 863  11.66666667  3 1.184326e-02
#> 864  11.75000000  3 1.147388e-02
#> 865  11.83333333  3 1.111602e-02
#> 866  11.91666667  3 1.076932e-02
#> 867  12.00000000  3 1.043344e-02
#> 868  12.08333333  3 1.010803e-02
#> 869  12.16666667  3 9.792773e-03
#> 870  12.25000000  3 9.487347e-03
#> 871  12.33333333  3 9.191447e-03
#> 872  12.41666667  3 8.904775e-03
#> 873  12.50000000  3 8.627044e-03
#> 874  12.58333333  3 8.357976e-03
#> 875  12.66666667  3 8.097299e-03
#> 876  12.75000000  3 7.844753e-03
#> 877  12.83333333  3 7.600083e-03
#> 878  12.91666667  3 7.363045e-03
#> 879  13.00000000  3 7.133399e-03
#> 880  13.08333333  3 6.910916e-03
#> 881  13.16666667  3 6.695371e-03
#> 882  13.25000000  3 6.486550e-03
#> 883  13.33333333  3 6.284241e-03
#> 884  13.41666667  3 6.088242e-03
#> 885  13.50000000  3 5.898356e-03
#> 886  13.58333333  3 5.714393e-03
#> 887  13.66666667  3 5.536167e-03
#> 888  13.75000000  3 5.363499e-03
#> 889  13.83333333  3 5.196217e-03
#> 890  13.91666667  3 5.034153e-03
#> 891  14.00000000  3 4.877143e-03
#> 892  14.08333333  3 4.725030e-03
#> 893  14.16666667  3 4.577661e-03
#> 894  14.25000000  3 4.434889e-03
#> 895  14.33333333  3 4.296569e-03
#> 896  14.41666667  3 4.162564e-03
#> 897  14.50000000  3 4.032737e-03
#> 898  14.58333333  3 3.906961e-03
#> 899  14.66666667  3 3.785107e-03
#> 900  14.75000000  3 3.667053e-03
#> 901  14.83333333  3 3.552681e-03
#> 902  14.91666667  3 3.441877e-03
#> 903  15.00000000  3 3.334529e-03
#> 904  15.08333333  3 3.230528e-03
#> 905  15.16666667  3 3.129771e-03
#> 906  15.25000000  3 3.032157e-03
#> 907  15.33333333  3 2.937587e-03
#> 908  15.41666667  3 2.845967e-03
#> 909  15.50000000  3 2.757204e-03
#> 910  15.58333333  3 2.671210e-03
#> 911  15.66666667  3 2.587898e-03
#> 912  15.75000000  3 2.507184e-03
#> 913  15.83333333  3 2.428987e-03
#> 914  15.91666667  3 2.353230e-03
#> 915  16.00000000  3 2.279835e-03
#> 916  16.08333333  3 2.208729e-03
#> 917  16.16666667  3 2.139841e-03
#> 918  16.25000000  3 2.073102e-03
#> 919  16.33333333  3 2.008444e-03
#> 920  16.41666667  3 1.945803e-03
#> 921  16.50000000  3 1.885115e-03
#> 922  16.58333333  3 1.826320e-03
#> 923  16.66666667  3 1.769359e-03
#> 924  16.75000000  3 1.714175e-03
#> 925  16.83333333  3 1.660712e-03
#> 926  16.91666667  3 1.608916e-03
#> 927  17.00000000  3 1.558735e-03
#> 928  17.08333333  3 1.510120e-03
#> 929  17.16666667  3 1.463021e-03
#> 930  17.25000000  3 1.417391e-03
#> 931  17.33333333  3 1.373184e-03
#> 932  17.41666667  3 1.330356e-03
#> 933  17.50000000  3 1.288863e-03
#> 934  17.58333333  3 1.248665e-03
#> 935  17.66666667  3 1.209720e-03
#> 936  17.75000000  3 1.171990e-03
#> 937  17.83333333  3 1.135437e-03
#> 938  17.91666667  3 1.100024e-03
#> 939  18.00000000  3 1.065716e-03
#> 940  18.08333333  3 1.032477e-03
#> 941  18.16666667  3 1.000275e-03
#> 942  18.25000000  3 9.690776e-04
#> 943  18.33333333  3 9.388531e-04
#> 944  18.41666667  3 9.095712e-04
#> 945  18.50000000  3 8.812027e-04
#> 946  18.58333333  3 8.537189e-04
#> 947  18.66666667  3 8.270923e-04
#> 948  18.75000000  3 8.012961e-04
#> 949  18.83333333  3 7.763045e-04
#> 950  18.91666667  3 7.520924e-04
#> 951  19.00000000  3 7.286354e-04
#> 952  19.08333333  3 7.059100e-04
#> 953  19.16666667  3 6.838934e-04
#> 954  19.25000000  3 6.625635e-04
#> 955  19.33333333  3 6.418989e-04
#> 956  19.41666667  3 6.218787e-04
#> 957  19.50000000  3 6.024830e-04
#> 958  19.58333333  3 5.836921e-04
#> 959  19.66666667  3 5.654874e-04
#> 960  19.75000000  3 5.478504e-04
#> 961  19.83333333  3 5.307635e-04
#> 962  19.91666667  3 5.142096e-04
#> 963  20.00000000  3 4.981719e-04
#> 964  20.08333333  3 4.826345e-04
#> 965  20.16666667  3 4.675816e-04
#> 966  20.25000000  3 4.529982e-04
#> 967  20.33333333  3 4.388697e-04
#> 968  20.41666667  3 4.251818e-04
#> 969  20.50000000  3 4.119208e-04
#> 970  20.58333333  3 3.990734e-04
#> 971  20.66666667  3 3.866267e-04
#> 972  20.75000000  3 3.745683e-04
#> 973  20.83333333  3 3.628859e-04
#> 974  20.91666667  3 3.515678e-04
#> 975  21.00000000  3 3.406028e-04
#> 976  21.08333333  3 3.299797e-04
#> 977  21.16666667  3 3.196880e-04
#> 978  21.25000000  3 3.097173e-04
#> 979  21.33333333  3 3.000575e-04
#> 980  21.41666667  3 2.906990e-04
#> 981  21.50000000  3 2.816324e-04
#> 982  21.58333333  3 2.728486e-04
#> 983  21.66666667  3 2.643388e-04
#> 984  21.75000000  3 2.560943e-04
#> 985  21.83333333  3 2.481070e-04
#> 986  21.91666667  3 2.403688e-04
#> 987  22.00000000  3 2.328719e-04
#> 988  22.08333333  3 2.256089e-04
#> 989  22.16666667  3 2.185724e-04
#> 990  22.25000000  3 2.117554e-04
#> 991  22.33333333  3 2.051509e-04
#> 992  22.41666667  3 1.987525e-04
#> 993  22.50000000  3 1.925536e-04
#> 994  22.58333333  3 1.865481e-04
#> 995  22.66666667  3 1.807298e-04
#> 996  22.75000000  3 1.750930e-04
#> 997  22.83333333  3 1.696321e-04
#> 998  22.91666667  3 1.643414e-04
#> 999  23.00000000  3 1.592158e-04
#> 1000 23.08333333  3 1.542500e-04
#> 1001 23.16666667  3 1.494391e-04
#> 1002 23.25000000  3 1.447783e-04
#> 1003 23.33333333  3 1.402628e-04
#> 1004 23.41666667  3 1.358881e-04
#> 1005 23.50000000  3 1.316499e-04
#> 1006 23.58333333  3 1.275439e-04
#> 1007 23.66666667  3 1.235659e-04
#> 1008 23.75000000  3 1.197120e-04
#> 1009 23.83333333  3 1.159784e-04
#> 1010 23.91666667  3 1.123611e-04
#> 1011 24.00000000  3 1.088567e-04
#> 1012 24.08333333  3 1.054616e-04
#> 1013 24.16666667  3 1.021723e-04
#> 1014 24.25000000  3 9.898567e-05
#> 1015 24.33333333  3 9.589841e-05
#> 1016 24.41666667  3 9.290744e-05
#> 1017 24.50000000  3 9.000975e-05
#> 1018 24.58333333  3 8.720244e-05
#> 1019 24.66666667  3 8.448269e-05
#> 1020 24.75000000  3 8.184776e-05
#> 1021 24.83333333  3 7.929502e-05
#> 1022 24.91666667  3 7.682189e-05
#> 1023 25.00000000  3 7.442589e-05
#> 1024 25.08333333  3 7.210463e-05
#> 1025 25.16666667  3 6.985576e-05
#> 1026 25.25000000  3 6.767703e-05
#> 1027 25.33333333  3 6.556625e-05
#> 1028 25.41666667  3 6.352131e-05
#> 1029 25.50000000  3 6.154015e-05
#> 1030 25.58333333  3 5.962078e-05
#> 1031 25.66666667  3 5.776127e-05
#> 1032 25.75000000  3 5.595975e-05
#> 1033 25.83333333  3 5.421443e-05
#> 1034 25.91666667  3 5.252353e-05
#> 1035 26.00000000  3 5.088538e-05
#> 1036 26.08333333  3 4.929832e-05
#> 1037 26.16666667  3 4.776075e-05
#> 1038 26.25000000  3 4.627115e-05
#> 1039 26.33333333  3 4.482800e-05
#> 1040 26.41666667  3 4.342986e-05
#> 1041 26.50000000  3 4.207533e-05
#> 1042 26.58333333  3 4.076304e-05
#> 1043 26.66666667  3 3.949168e-05
#> 1044 26.75000000  3 3.825998e-05
#> 1045 26.83333333  3 3.706669e-05
#> 1046 26.91666667  3 3.591062e-05
#> 1047 27.00000000  3 3.479061e-05
#> 1048 27.08333333  3 3.370552e-05
#> 1049 27.16666667  3 3.265428e-05
#> 1050 27.25000000  3 3.163583e-05
#> 1051 27.33333333  3 3.064914e-05
#> 1052 27.41666667  3 2.969323e-05
#> 1053 27.50000000  3 2.876713e-05
#> 1054 27.58333333  3 2.786991e-05
#> 1055 27.66666667  3 2.700068e-05
#> 1056 27.75000000  3 2.615855e-05
#> 1057 27.83333333  3 2.534269e-05
#> 1058 27.91666667  3 2.455228e-05
#> 1059 28.00000000  3 2.378652e-05
#> 1060 28.08333333  3 2.304465e-05
#> 1061 28.16666667  3 2.232591e-05
#> 1062 28.25000000  3 2.162958e-05
#> 1063 28.33333333  3 2.095498e-05
#> 1064 28.41666667  3 2.030142e-05
#> 1065 28.50000000  3 1.966824e-05
#> 1066 28.58333333  3 1.905480e-05
#> 1067 28.66666667  3 1.846051e-05
#> 1068 28.75000000  3 1.788474e-05
#> 1069 28.83333333  3 1.732693e-05
#> 1070 28.91666667  3 1.678653e-05
#> 1071 29.00000000  3 1.626297e-05
#> 1072 29.08333333  3 1.575575e-05
#> 1073 29.16666667  3 1.526434e-05
#> 1074 29.25000000  3 1.478826e-05
#> 1075 29.33333333  3 1.432703e-05
#> 1076 29.41666667  3 1.388019e-05
#> 1077 29.50000000  3 1.344728e-05
#> 1078 29.58333333  3 1.302787e-05
#> 1079 29.66666667  3 1.262155e-05
#> 1080 29.75000000  3 1.222789e-05
#> 1081 29.83333333  3 1.184652e-05
#> 1082 29.91666667  3 1.147704e-05
#> 1083 30.00000000  3 1.111908e-05